requirements.txt
plaid-python <=37.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
requests>=2.31.0
boto3<=1.41.1
#dbt-core <=1.10.15
#snowflake <=1.9.0
apache-airflow<=3.1.3
dbt-bigquery
dbt-postgres
dbt-snowflake
dbt-databricks
dbt-redshift
dbt-athena



